refactor(#162): relocate sheet constants/utils to _core (P3a, prereq)
Prerequisite for the sheet.py split: move the page/scale-selection and
sheet-layout primitives that the compose-then-pack cluster needs — _DIM_PAD,
_STRIP_GAP, _ISO_WIDTH_BUDGET, _ISO_MIN_FIT_FRAC, _PAGE_SIZES, _SCALES, _LADDER,
and the page-parsing helpers _parse_page/_tb_width — from make_drawing.py to
_core.py. They are shared by the moving cluster (choose_scale/_layout_geometry,
P3b) and the staying repack pass (_repack_candidates uses _parse_page/_SCALES/
_LADDER), so they belong in the shared base; this keeps the DAG acyclic when
sheet.py lands.

Pure move: make_drawing re-imports all nine; _parse_page/_tb_width depend only on
_PAGE_SIZES + re (both in _core). Golden gate passes against the committed
snapshots with no regeneration; scale/page tests pass; ruff check + format + mypy
clean. Adversarially checked: _core imports standalone, _parse_page/_tb_width
behave identically (incl. the bad-page ValueError), no stale defs, the module-level
_DIM_PAD invariant assert still holds at import.
